deployement_recommendation-system-main/deployement_recommendation-system-main/requirement.txt
python-dotenv==0.19.2
sklearn==0.0
numpy
requests
joblib
nltk
urllib3
